jsabol/react-scripts Add support for nested directory structure
This adds a --dir flag to react-scripts to specify the subdirectory
where the react app is found.

For example, if your directory structure looks like this:
```
my-app
├── client
│ └── public
│ └── src
├── server
├── node_modules
├── package.json
```

Typically, you would need a separate package.json inside the /client
folder. With this update you should be able to have a single
package.json at the top-level directory and update your package.json
like this:

```diff
"scripts": {
...
- "start": "react-scripts start",
+ "start": "react-scripts start --dir client",
...
}
```